+++ src/pages/progress/components/DownloadProgressButton.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import GoogleAnalytics from "react-ga4";
 import formatProgressFileDownloadName from "../utils/formatProgressFileDownloadName";
-import makeDownloadHref from "../utils/makeDownloadHref";
+import makeDownloadHref from "../../../utils/makeDownloadHref";
 import type { MetWords } from "../../../types";
 
 type Props = {
