+++ src/pages/progress/components/ReformatProgress.tsx
@@ -1,7 +1,7 @@
 import React, { useCallback, useMemo, useState } from "react";
 import GoogleAnalytics from "react-ga4";
 import formatSpacePlacementValue from "../utils/formatSpacePlacementValue";
-import makeDownloadHref from "../utils/makeDownloadHref";
+import makeDownloadHref from "../../../utils/makeDownloadHref";
 import trimAndSumUniqMetWords from "../../../utils/trimAndSumUniqMetWords";
 import formatProgressFileDownloadName from "../utils/formatProgressFileDownloadName";
 import type { MetWords, UserSettings } from "../../../types";
